Actualización del sistema: nuevas funcionalidades y mejoras

+++ database/seeders/DatabaseSeeder.php
@@ -2,17 +2,16 @@
 
 namespace Database\Seeders;
 
-// use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
+
 
 class DatabaseSeeder extends Seeder
 {
-    /**
-     * Seed the application's database.
-     */
     public function run(): void
     {
+        // Primero vaciamos las tablas
         $this->truncateTables([
             // 'role_user',
             'users',
@@ -31,7 +30,7 @@ class DatabaseSeeder extends Seeder
             'ingresos',
             'detalle_ingresos',
             'producto_sucursals',
-            'tipo_pagos',
+            // 'tipo_pagos',
             'clientes',
             'proformas',
             'detalle_proformas',
@@ -46,27 +45,26 @@ class DatabaseSeeder extends Seeder
             'videos',
             'configs',
         ]);
-        // \App\Models\User::factory(10)->create();
 
-        // \App\Models\User::factory()->create([
-        //     'name' => 'Test User',
-        //     'email' => 'test@example.com',
-        // ]);
+        // Luego insertamos los datos
         $this->call([
+            VentasPrediccionesSeeder::class,
             ConfigurationSeeder::class,
             // PruebaSeeder::class,
-
         ]);
     }
 
-    public function truncateTables(array $tables)
-    {
-        DB::statement('SET FOREIGN_KEY_CHECKS = 0;');
+   public function truncateTables(array $tables)
+{
+    DB::statement('SET FOREIGN_KEY_CHECKS = 0;');
 
-        foreach ($tables as $table) {
+    foreach ($tables as $table) {
+        if (Schema::hasTable($table)) {
             DB::table($table)->truncate();
         }
-
-        DB::statement('SET FOREIGN_KEY_CHECKS = 1;');
     }
+
+    DB::statement('SET FOREIGN_KEY_CHECKS = 1;');
+}
+
 }
